# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Math With Hawra API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quizzes/{id}/submissions:
    get:
      operationId: getQuizSubmissions
      tags: [submissions]
      summary: Get all submissions for a quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submissions grouped by student
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    delete:
      operationId: deleteQuizSubmissions
      tags: [submissions]
      summary: Delete selected or all submissions for a quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteSubmissionsInput"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /submissions:
    post:
      operationId: submitQuiz
      tags: [submissions]
      summary: Student submits quiz answers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResult"
  /submissions/{id}/answers/{answerId}/grade:
    patch:
      operationId: gradeAnswer
      tags: [submissions]
      summary: Teacher grades an essay answer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: answerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeInput"
      responses:
        "200":
          description: Graded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    QuizOption:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        imageUrl:
          type: ["string", "null"]
        order:
          type: integer
      required: [id, text, order]
...
    AnswerDetail:
      type: object
      properties:
        id:
          type: integer
        questionId:
          type: integer
        questionText:
          type: string
        questionType:
          type: string
        answerText:
          type: ["string", "null"]
        answerIndex:
          type: ["integer", "null"]
        isCorrect:
          type: ["boolean", "null"]
        grade:
          type: ["integer", "null"]
        status:
          type: string
          enum: [auto, pending, graded]
        modelAnswer:
          type: ["string", "null"]
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuizOption"
      required: [id, questionId, questionText, questionType, status]
    Submission:
      type: object
      properties:
        id:
          type: integer
        quizId:
          type: integer
        quizTitle:
          type: string
        studentName:
          type: string
        score:
          type: ["integer", "null"]
        earnedPoints:
          type: integer
        totalPoints:
          type: integer
        completedAt:
          type: string
        hasPending:
          type: boolean
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerDetail"
      required: [id, quizId, quizTitle, studentName, earnedPoints, totalPoints, completedAt, hasPending]
    AnswerInput:
      type: object
      properties:
        questionId:
          type: integer
        answerText:
          type: ["string", "null"]
        answerIndex:
          type: ["integer", "null"]
      required: [questionId]
    SubmissionInput:
      type: object
      properties:
        quizId:
          type: integer
        studentName:
          type: string
        timeSpentSeconds:
          type: integer
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerInput"
      required: [quizId, studentName, timeSpentSeconds, answers]
    SubmissionResult:
      type: object
      properties:
        id:
          type: integer
        score:
          type: ["integer", "null"]
        earnedPoints:
          type: integer
        totalPoints:
          type: integer
        hasPending:
          type: boolean
      required: [id, earnedPoints, totalPoints, hasPending]
    GradeInput:
      type: object
      properties:
        grade:
          type: integer
          minimum: 0
          maximum: 100
      required: [grade]
    DeleteSubmissionsInput:
      type: object
      properties:
        submissionIds:
          type: array
          items:
            type: integer
        deleteAll:
          type: boolean
      required: []
...