# lib/api-zod/src/generated/types/answerInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Math With Hawra API
 * OpenAPI spec version: 0.1.0
 */

export interface AnswerInput {
  questionId: number;
  /** @nullable */
  answerText?: string | null;
  /** @nullable */
  answerIndex?: number | null;
}