# lib/api-zod/src/generated/types/submissionResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Math With Hawra API
 * OpenAPI spec version: 0.1.0
 */

export interface SubmissionResult {
  id: number;
  /** @nullable */
  score?: number | null;
  earnedPoints: number;
  totalPoints: number;
  hasPending: boolean;
}